fix: remaining Copilot review items — plist revert, dismiss overlap, test structure
- Revert xcschememanagement.plist to main (user-specific, should not be in PR)
- Add end padding to reply preview Row so 48dp dismiss button doesn't
  overlap thumbnail/event chip
- Extract shared test helpers from UnifiedInputBubbleContractsTest into
  dedicated InputBarContractHelpers.kt — eliminates hidden cross-file dependency
- Replace duplicate SIGNAL_TRANSITION_DURATION_MS with production
  SIGNAL_TRANSITION_MS constant
